add worker.h
Move struct worker_state and connection into worker.h.

These will be used by a future iou.c that implement io_uring.
